requirements.txt
beautifulsoup4
Unidecode
python-dotenv
supabase